+++ toolchain/parser/function.go
@@ -132,49 +132,3 @@ func parseArgument(p *Parser) *ast.Argument {
 		Type: p.parseType(),
 	}
 }
-
-// Parses a function call that can have a form:
-// <name>(<arguments>)
-func (p *Parser) parseFunctionCall(left ast.Node) *ast.FunctionCall {
-	p.Consume()
-
-	var path *ast.Path
-	if left.Kind() == ast.IdentifierNode {
-		path = &ast.Path{
-			Segments: []ast.IdentifierLiteral{*left.(*ast.IdentifierLiteral)},
-		}
-	} else if left.Kind() == ast.PathNode {
-		path = left.(*ast.Path)
-	} else {
-		dbg := debug.NewSourceLocation(p.Source(), left.Location().Row, left.Location().Column)
-		dbg.ThrowError("Expected identifier or path", true, debug.NewHint("Did you forget to add a function name?", "fn"))
-	}
-
-	args := make([]ast.Node, 0)
-
-	for p.Current().Kind != tokens.RightParenthesis {
-		arg := p.parseExpression()
-		if arg == nil {
-			break
-		}
-
-		args = append(args, arg)
-
-		if p.Current().Kind == tokens.Comma {
-			p.Consume()
-		}
-	}
-
-	last := p.Consume()
-
-	if last.Kind != tokens.RightParenthesis {
-		dbg := debug.NewSourceLocation(p.Source(), left.Location().Row, left.Location().Column)
-		dbg.ThrowError("Expected closing parenthesis", true, debug.NewHint("Did you forget to add a closing parenthesis?", ")"))
-	}
-
-	return &ast.FunctionCall{
-		Path: *path,
-		Args: args,
-		Loc:  last.Location,
-	}
-}
